moved pipeline options
